All Test Cases added for 02-01-01

+++ Unit-02/Chapter-01/02-01-01.py
@@ -1,2 +1,2 @@
-a = int(input())
-print(a > 100)
+a = float(input())
+print(a > 100.0)
